Backend: added indicator combinations

+++ backend/src/objects/indicators/implementations/bollinger_bands.rs
@@ -30,13 +30,13 @@ impl BollingerBands {
                 IndicatorParameter {
                     name: "period".to_string(),
                     description: "The number of periods to use in the calculation.".to_string(),
-                    r#type: "i32".to_string(),
+                    r#type: "integer".to_string(),
                     default: "20".to_string(),
                 },
                 IndicatorParameter {
                     name: "deviation".to_string(),
                     description: "The number of standard deviations to use in the calculation.".to_string(),
-                    r#type: "f64".to_string(),
+                    r#type: "float".to_string(),
                     default: "2.0".to_string(),
                 },
             ],
